chore(common): move DevWorkspaceStatus enum to @eclipse-che/common
Remove the backend-local copy (devWorkspaceStatus.ts) and the frontend
inline definition. Both now import from @eclipse-che/common so the
enum has a single canonical source.
