--- packages/dashboard-backend/src/devworkspaceClient/devWorkspaceStatus.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright (c) 2018-2025 Red Hat, Inc.
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- */
-
-export enum DevWorkspaceStatus {
-  FAILED = 'Failed',
-  FAILING = 'Failing',
-  STARTING = 'Starting',
-  TERMINATING = 'Terminating',
-  RUNNING = 'Running',
-  STOPPED = 'Stopped',
-  STOPPING = 'Stopping',
-}
